change ncs to mef
